+++ R/trakt.getSeasons.R
@@ -21,9 +21,6 @@
 #' breakingbad.seasons <- trakt.getSeasons("breaking-bad", extended = "min")
 #' }
 trakt.getSeasons <- function(target, extended = "full,images", dropspecials = TRUE){
-  if (is.null(getOption("trakt.headers"))){
-    stop("HTTP headers not set, see ?get_trakt_credentials")
-  }
 
   # Construct URL, make API call
   baseURL <- "https://api-v2launch.trakt.tv/shows/"