+++ R/trakt.movie.people.R
@@ -17,9 +17,6 @@
 #' tron.people <- trakt.movie.people("tron-legacy-2010")
 #' }
 trakt.movie.people <- function(target, extended = "min"){
-  if (is.null(getOption("trakt.headers"))){
-    stop("HTTP headers not set, see ?get_trakt_credentials")
-  }
 
   # Construct URL, make API call
   baseURL  <- "https://api-v2launch.trakt.tv/movies"