+++ R/trakt.user.collection.R
@@ -16,9 +16,6 @@
 #' seans.movies <- trakt.user.collection(user = "sean", type = "movies)
 #' }
 trakt.user.collection <- function(user = getOption("trakt.username"), type = "shows"){
-  if (is.null(getOption("trakt.headers"))){
-    stop("HTTP headers not set, see ?get_trakt_credentials")
-  }
   if (is.null(user) && is.null(getOption("trakt.username"))){
     stop("No username is set.")
   }