+++ R/trakt.user.following.R
@@ -13,9 +13,6 @@
 #' trakt.user.following("jemus42")
 #' }
 trakt.user.following <- function(user = getOption("trakt.username")){
-  if (is.null(getOption("trakt.headers"))){
-    stop("HTTP headers not set, see ?get_trakt_credentials")
-  }
   if (is.null(user) && is.null(getOption("trakt.username"))){
     stop("No username is set.")
   }