+++ R/trakt.user.stats.R
@@ -14,9 +14,6 @@
 #' seanstats <- trakt.user.stats(user = "sean")
 #' }
 trakt.user.stats <- function(user = getOption("trakt.username")){
-  if (is.null(getOption("trakt.headers"))){
-    stop("HTTP headers not set, see ?get_trakt_credentials")
-  }
   if (is.null(user) && is.null(getOption("trakt.username"))){
     stop("No username is set.")
   }